Used the function `normalize_data_format` to remove some code. Moved the implementation to keras/backend/common.py.
