+++ docs/autogen.py
@@ -99,6 +99,7 @@ EXCLUDE = {
     'deserialize',
     'get',
     'set_image_dim_ordering',
+    'normalize_data_format',
     'image_dim_ordering',
     'get_variable_shape',
 }
